Move survey routes into survey scope
