+++ app/views/course/survey/surveys/_survey.json.jbuilder
