+++ app/views/course/survey/surveys/index.html.slim
