+++ app/views/course/survey/surveys/index.json.jbuilder
