+++ app/views/course/survey/surveys/show.json.jbuilder
