+++ client/app/bundles/course/survey/components/AddButton.jsx
