+++ client/app/bundles/course/survey/components/SurveyDetails.jsx
