+++ client/app/bundles/course/survey/constants.jsx
