+++ client/app/bundles/course/survey/containers/NewSurveyButton.jsx
