+++ client/app/bundles/course/survey/reducers/canCreate.jsx
