+++ client/app/bundles/course/survey/reducers/deleteConfirmation.jsx
