+++ client/app/bundles/course/survey/reducers/index.jsx
