+++ client/app/bundles/course/survey/reducers/notification.jsx
