+++ client/app/bundles/course/survey/reducers/surveys.jsx
