+++ client/app/bundles/course/survey/surveys.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { render } from 'react-dom';
 import { Router, browserHistory } from 'react-router';
 import ProviderWrapper from 'lib/components/ProviderWrapper';
-import storeCreator from './surveys/store';
-import routes from './surveys/routes';
+import storeCreator from './store';
+import routes from './routes';
 
 const mountNode = document.getElementById('course-survey-component');
