+++ client/app/bundles/course/survey/translations.jsx
